requirements.txt
colorama==0.4.4
Cython==0.29.24
h5py==3.4.0
imageio==2.9.0
jsonschema==3.2.0
kornia==0.5.10
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.2
Pillow==8.3.1
prettytable==2.2.0
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
pyfhog @ git+https://github.com/JBamberger/pyfhog.git@e36612305e2a18ac5b9ab418ffb80db9cc31c6ef
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
spatial-correlation-sampler==0.3.0
tensorboard==2.6.0
timm==0.4.12
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.2
